ocr-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pytesseract==0.3.13
Pillow==11.0.0
opencv-python==4.10.0.84
numpy==2.1.1

